Mobile view enhancements
