+++ app/(protected)/layout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { Sidebar } from "../components/Sidebar";
+import { Toolbar } from "../components/Toolbar";
 
 type Props = {
   children: ReactNode;
@@ -9,7 +9,7 @@ export default function ProtectedLayout({ children }: Props) {
   return (
     <div className="flex flex-col-reverse md:flex-row gap-4 max-w-[1640px] mx-auto">
       <div className="fixed bottom-0 w-full z-10 md:relative md:w-fit">
-        <Sidebar />
+        <Toolbar />
       </div>
       {children}
     </div>
